flight: big code refactor
1. slim down openpilot.h -- getting closer to eliminate this.
2. factor some startup code to chibi_main.c from individual targets,
as part of ongoing work to simplify the target files.
3. include taskmonitor.h from pios_thread, since they're pretty much
always used together.
4. adjust orderings in taskmonitor.h/pios.thread so that idempotency
works as intended here.
5. include pios_thread.h several additional places.
