+++ flight/Modules/VTXConfig/VTXConfig.c
@@ -33,6 +33,7 @@
 
 #include "openpilot.h"
 #include <pios_hal.h>
+#include <pios_thread.h>
 
 #include "tbs_smartaudio.h"
