+++ flight/Modules/VtolPathFollower/vtol_follower_fsm.c
@@ -46,6 +46,8 @@
 
 #include "openpilot.h"
 
+#include <pios_thread.h>
+
 #include "coordinate_conversions.h"
 #include "physical_constants.h"
