+++ flight/UAVObjects/inc/uavobjecttemplate.h
@@ -9,7 +9,7 @@
  * @file       $(NAMELC).h
  * @author     The OpenPilot Team, http://www.openpilot.org Copyright (C) 2010.
  * @author     Tau Labs, http://taulabs.org, Copyright (C) 2012-2015
- * @author     dRonin, http://dronin.org Copyright (C) 2015
+ * @author     dRonin, http://dronin.org Copyright (C) 2015-2017
  * @brief      Include files for the $(NAME) object. This file has been 
  *             automatically generated by the UAVObjectGenerator.
  * 
@@ -39,6 +39,7 @@
 #define $(NAMEUC)_H
 
 #include "pios_queue.h"
+#include "uavobjectmanager.h"
 #include "uavoversion.h"
 
 $(PARENT_INCLUDES)
