+++ flight/targets/dtfc/fw/pios_board.c
@@ -71,10 +71,6 @@ uintptr_t pios_com_openlog_logging_id;
 
 void PIOS_Board_Init(void)
 {
-
-	/* Delay system */
-	PIOS_DELAY_Init();
-
 	const struct pios_board_info *bdinfo = &pios_board_info_blob;
 
 #if defined(PIOS_INCLUDE_ANNUNC)
@@ -107,14 +103,6 @@ void PIOS_Board_Init(void)
 		PIOS_HAL_CriticalError(PIOS_LED_ALARM, PIOS_HAL_PANIC_FILESYS);
 #endif	/* PIOS_INCLUDE_FLASH */
 
-	/* Initialize the task monitor library */
-	TaskMonitorInitialize();
-
-	/* Initialize UAVObject libraries */
-	UAVObjInitialize();
-
-	/* Initialize the alarms library. Reads RCC reset flags */
-	AlarmsInitialize();
 	PIOS_RESET_Clear(); // Clear the RCC reset flags after use.
 
 	/* Initialize the hardware UAVOs */
